The corrects have been added

+++ src/main/resources/liquibase/scripts/notification_task_creator.sql
@@ -2,8 +2,8 @@
 
 -- changeset q100s:1
 CREATE TABLE notification_task (
-    id SERIAL PRIMARY KEY,
-    chat_id SERIAL NOT NULL,
+    id bigserial PRIMARY KEY,
+    chat_id bigint NOT NULL,
     notification_text TEXT NOT NULL,
     notification_time TIMESTAMP NOT NULL
 );
